#165: CenterNavigationHeaderView 생성
